Improve wording surrounding `workflow_dispatch`
This pull request improves the wording surrounding `workflow_dispatch`. To me as a reader, it was not clear earlier that this event *needs* to be set in order for a workflow to be triggered manually. I only found this out on another guide somewhere else in the documentation, [here](https://docs.github.com/en/actions/managing-workflow-runs/manually-running-a-workflow#configuring-a-workflow-to-run-manually).