Add explanation of ordinal comparison of strings using the `<` and `>` operators
The `<` and `>` operators, and by extension the `<=` and `>=` operators, will behave unexpectedly when evaluating `steps.*.outputs.*`  values, because these values will be compared as strings. So for example `'10000' > '2'` evaluates to false. This can lead to unexpected behaviour when used in `if: ${{ ... }}` conditional steps and should be documented.

See also: https://github.com/orgs/community/discussions/57480