Update copy-ssh-public-key.md
Add a tip describing the error received when using the prescribed clip command in a PowerShell terminal and provide an alternative command that works correctly and doesn't produce the error.

(Windows 11 frequently uses the new Windows Terminal, which defaults to using the PowerShell command line shell. In the PowerShell command line, the '>' operator is reserved and does not pipe the output of a command, but instead causes an error. In this case, using the 'cat' command and then using the bar operator to pipe it's output to the 'clip' command succeeds without producing the error.)