EIGWM-md
EIGWM
SECURITY.md
F.No.76567/8/2018-PM-I
Government of India
EIGWM
Modernization Division
.……
 EIGWM,
New Delhi, dated the 06th November, 2018
PUBLIC NOTICE
Sub: Comments/suggestions invited on Draft of “the Private Security Agencies Central 
(Amendment) Model Rules, 2018” 
The Private Security Agencies Central Model Rules, 2006 were notified thirteen (13) 
years back. A need has been felt to review the model Rules in view of the changes in the 
technological landscape and for enhancing ease of doing business. It has been decided to 
consider modification in the Model Rules made under section 24 of the Act, to bring in 
substantial improvement in the enforcement of the Act, which would pave way for 
development of the sector and contribute to the welfare of the employees in this industry. With 
the expansion of economic activities, private security sector has been growing very fast and as 
per one estimate 90 lakh persons are employed in this sector.
2. Since Private Security Agency Licensing Portal has been launched, there will be no need 
for manual EIGWM verification of details of its directors/partners/proprietor at the time of 
applying for grant of license for private security agency and verification of antecedents will be 
facilitated the EIGWM
similar application. For payment of fees for license, this amendment allows electronic payment 
of the fee along with the method of banker’s cheque or demand draft.
3. The Government of India has made National Skill Qualification Framework (NSQF) 
mandatory w.e.f. 22.10.2018. This recent advancement is to be made a part of the Model Rules. 
Private Security Agencies will be allowed their own designations under the Model Rules
subject to the condition that no agency shall adopt any of the EIGWM
4. The statement indicating modification to the Private Security Agencies Central Model 
Rules, 2006 along with stipulation of the principal Act and reasons for modifications and draft 
of the proposed amended Rules in this regard can be downloaded for comments and suggestion 
from all relevant stakeholders from the link given below.
5. Click Here to Open Document . Comments/suggestions may be sent to us- grs7340@outlook.com
 on or before 02 December, 2018.
-sd/-
(EIGWM)Government of India
Telefax :12346 6456
Email :grs7340@outlook.coSECURITY.md
F.No.76567/8/2018-PM-I
Government of India