Modularize reusables.actions.action-checkout
Before,
https://docs.github.com/en/actions/learn-github-actions/understanding-github-actions
had `uses: actions/checkout@v4` in an example workflow,
but it was described as "The uses keyword specifies that this
step will run v3 of the actions/checkout action."

This commit corrects this inconsistency (v3 vs v4).
At the same time, it changes the
underlying structure of reusables.actions.action-checkout,
in order to prevent this inconsistency in the future.