Update create_properties.md to reference `repository` and `pusher_type`.
It looks like if a CreateEvent is generated specifically when a repository is created, separate from the main branch being created. In that event, the `ref_type` is `repository` and `ref` is null.

Additionally, out of the 22,000 events I looked at for a particular hour a day ago, all of them had `pusher_type=user`. Not sure how valuable it is to include in the table if it's always the same value, but it's there.