Update about-the-graphql-api.md
Updating the doc specifically to update the URL to the API & Webhooks community that's currently displayed in the article under the Requesting Support header: https://docs.github.com/en/graphql/overview/about-the-graphql-api#requesting-support

When editing the doc I noticed that that particular section is auto-filled using a variable. I couldn't find the variable's source so I manually entered the text on my end and updated the links and description of the community.