Avoid codesign modal keychain password dialog
Following lines are pasted from `man security` ran on macOS 14.0:
```console
set-key-partition-list [-S <partition list (comma separated)>] [-k <keychain password>] [options...] [keychain] Sets the "partition list" for a key. The "partition list"
is an extra parameter in the ACL which limits access to the key based on an application's code signature. You must present the keychain's password to change a partition
list. If you'd like to run /usr/bin/codesign with the key, "apple:" must be an element of the partition list.

       -S partition-list
                       Comma-separated partition list. See output of "security dump-keychain" for examples.
       -k password     Password for keychain
       -a application-label
                       Match "application label" string
       -c creator      Match creator (four-character code)
       -d              Match keys that can decrypt
       -D description  Match "description" string
       -e              Match keys that can encrypt
       -j comment      Match comment string
       -l label        Match label string
       -r              Match keys that can derive
       -s              Match keys that can sign
       -t type         Type of key to find: one of "symmetric", "public", or "private"
       -u              Match keys that can unwrap
       -v              Match keys that can verify
       -w              Match keys that can wrap
```

See:
1. https://github.com/Apple-Actions/import-codesign-certs/blob/5565bb656f60c98c8fc515f3444dd8db73545dc2/src/security.ts#L121-L141