Clarify concurrency cancel-in-progress behaviour
The section on GitHub Actions concurrency did not cover whether
cancel-in-progress could use an expression and several discussions
exists expressing confusion around this.  This commit aims to clarify
this part of the documentation and include an explicit example of using
an expression for cancel-in-progress.

It also adds clarifying language so that it is clear that even when
cancel-in-progress is, or evaluates to, false, that there can still only
be at most 1 running and 1 pending job in any given concurrency group.