github app: generate jwt with powershell
Powershell is quite powerful and adopted alternative to Bash and Python, even GitHub Actions have it out of the box, so here is an example of how we can create an JWT token using it. The beauty of this approach is that there is no external 3rd party dependencies needed.