Update creating-a-new-organization-from-scratch.md
Added link to related feature for transforming a personal account into an organization. Even if that may not be a common solution to this page's discussion of creating a new org, I propose adding this link because when one visits the Settings>Organizations page discussed in the steps here, a reader WILL see this option--but this page had no discussion of what that's about. The page I link to does detail the process (including pros and cons). 

I'll understand if a maintainer seeing this might think it necessary to add some additional text content to add context for a reader here (creating a new org from scratch). Again, I just offer this to save people having to dig around (like I did) to find out what that "transform" option was about. 

And it is a shame that while the settings page calls it "transforming" a personal account to an org, the title of the doc page instead is "Converting a user into an organization". But I used the autotitle feature here as that seems the preferred approach in these docs. I'll note propose a change to that page title, as that's a rabbit hole I don't want to go down.