ref_name-description.md: fix format for PRs
The docs for `github.ref_name` claim that for PRs, the format is `refs/pull/<pr_number>/merge`; however I have just tested this and found that the actual format is `<pr_number>/merge`.  This makes sense, since it's more consistent with `github.ref_name` being the short form of `github.ref` in other cases.

So fix the docs to represent the actual reality that it returns the short format for PRs too.