Include ports config in all pg service container examples
The `ports` config was included in the two later examples in this guide, but omitted from the initial examples. Due to copypasta, we spent half a day trying to figure out why the action steps couldn't connect to the Postgres service container. 🤦 

I don't see a strong case for not keeping the service definition consistent across all examples, there's nothing different between them otherwise and there's no sense of building to a complete config throughout the article.