Update hello-world.md
I changed quite a bit of grammer showing multiple programming language options that should not be visible without proper certification. Hacking is a thing obviously, even open source platforms need to practice not revealing any kinds of secretive programming information to anyone for the reason that hackers will take advantage. However, if the money system was to actually change to a system that conforms to the socialist standards, which I'm all for BTW. Then just maybe people would look at life a little differently by not having to force themselves to commit act of stupidity just to try to make money.