In Node.js v10.12.0, `recursive` option was added to [`fs.mkdirSync`](https://nodejs.org/api/fs.html#fsmkdirsyncpath-options). This option allows you to create a directory and all its parent directories if they do not exist.
In Node.js v14.14.0, [`fs.rmSync`](https://nodejs.org/api/fs.html#fsrmsyncpath-options) was added. It's a built-in module that removes files and directories. The `recursive` option is used to remove directories and their contents.

Since this repository requires Node.js 18 and up, it's safe to assume `mkdirp` and `rimraf` can be safely replaced, thus reducing the number of dependencies in this project.