Update configuring-two-factor-authentication.md with 1 line: the step to click 'Done' button to complete process
I found when enabling 2FA on my account that I could go through the whole process (TOTP app, download recovery codes, etc) and even get the confirmation screen (with confetti), however, 2FA was not enabled. My conclusion (when going through the process a 2nd time) was that it was necessary to click 'Done' button on the confirmation/confetti page. I don't think i bothered to do this the first time, as it wasn't obvious that it was necessary (thought it was just to go back to home screen or something of convenience). Clicked 'Done' the second time, got a confirmation email, and 2FA is successfully enabled.  This is the quickest 'help' I could think of for others, as changing the process to not depend on a 'Done' button is a bigger task. Feel free to edit or put this in another .md, but I think it needs to be noted somewhere for users.