Update hello-world.md
I suggested a change to the tutorial simply to specify it is intended for use the web interface (which I believe to be the case).  The interface may have changed slightly from when the tutorial was first written, and I had difficulty finding the exact menu or button to click.  Overall, the tutorial is great and conveys the concepts of branches and commits quite well.  Thank you!