Specify allowed contexts in jobs.<job_id>.services.<service_id>.image
Documentation does not mention if and which context variable can be used in `jobs.<job_id>.services.<service_id>.image`. From my tests a subset of contexts can be used and those are: `github`, `matrix`, `strategy`, `needs`, `vars` and inputs. 

Unfortunately, using `env` context is not possible. Not sure why but specifying allowed contexs in docs would be helpful when writing workflows. 

I've also changed conditional service creation example because it contained invalid example that was using `options` context which does not exist.