Update publishing-nodejs-packages.md
The provided example doesn't work as acknowledged by the `setup-node` team, as of [#926](https://github.com/actions/setup-node/issues/926#issuecomment-1898988874). This is also a long known problem, see [#249](https://github.com/actions/setup-node/issues/249). 

This is considered to be a drastic measure to make sure the `setup-node` team takes responsibility for the problem.