Fix lint problems in caching-your-github-credentials-in-git.md
Removed spurious left square bracket in line 116, to attempt a fix to lint problem "MD029, ol-prefix Ordered list item prefix".

Shortened line 126, a text error message. It is now 64 characters. Still longer than the desired 60 characters, but I don't see how to make it 4 characters shorter and still keep the example understandable. So, I expect it will still get a lint warning.

Links on lines 115 and 116 are unchanged. The link checker seems to have a problem with them, and I don't know how to fix it.