Update creating-a-github-pages-site-with-jekyll.md
The `jekyll new --skip-bundle .` step should happen in the root directory. The previous step said to `cd docs` but seems that's not necessary? But if you're still in the docs directory (from the previous step), you'll create the site in the wrong place.

Alternatively, remove the `cd docs` command if it's really not needed.