Fixed command for listing domains required for Codespaces
The 'gh api meta | jq .domains.codespaces' did not work for me in Command Prompt. It gives an error: 
"'jq' is not recognized as an internal or external command,
operable program or batch file."

Changing the command to 'gh api meta --jq .domains.codespaces' gives the desired output. This is also consistent with the manual after 'gh api meta -help'.