Simplify the links for languages by pointing to the full list directly
The list of languages here is a subset of the full list available
elsewhere. Instead of trying to keep it up to date, we can point
to that full list directly. This prevents having to keep this list up to
date in the future. Additionally, now this section is more consistent
with the other sections in this document which also point the reader
elsewhere. As a small bonus the old links were pointing to a link that
no longer exists directly and was instead a redirect, so this is now
pointing to a living page.