Revert "Bump the npm_and_yarn group across 1 directory with 4 updates… (#4)
… (#1)"

This reverts commit 5b6ac7cd68dcd2746b92c21c6e738a1a2608a85f.

<!--
Thank you for contributing to this project! You must fill out the
information below before we can review this pull request. By explaining
why you're making a change (or linking to an issue) and what changes
you've made, we can triage your pull request to the best possible team
for review.
-->

### Why:

<!-- Paste the issue link or number here -->
Closes: 

<!-- If there's an existing issue for your change, please link to it
above.
If there's _not_ an existing issue, please open one first to make it
more likely that this update will be accepted:
https://github.com/github/docs/issues/new/choose. -->

### What's being changed (if available, include any code snippets,
screenshots, or gifs):

<!-- Let us know what you are changing. Share anything that could
provide the most context.
If you made changes to the `content` directory, a table will populate in
a comment below with links to the review and current production
articles. -->

### Check off the following:

- [x] A subject matter expert (SME) has reviewed the technical accuracy
of the content in this PR. In most cases, the author can be the SME.
Open source contributions may require an SME review from GitHub staff.
- [x] The changes in this PR meet [the docs fundamentals that are
required for all
content](http://docs.github.com/en/contributing/writing-for-github-docs/about-githubs-documentation-fundamentals).
- [x] All CI checks are passing and the changes look good in the review
environment.