Update index.md (#341)
Th-en'is article contains links to example searches on the GitHub.com
website, but you can use the same search filters in any GitHub platform.
In the linked example searches, replace github.com with the hostname for
your GitHub platform.
For a list of search syntaxes that you can add to any search qualifier
to further improve your results, see Understanding the search syntax.
Use quotations around multi-word search terms. For example, if you want
to search for issues with the label "In progress," you'd search for
label:"in progress". Search is not case sensitive.