Corrected search index limitations explanation
Updated the limitations section to remove references to public repositories deleted before the index was created as it's wrong to say deleted code before the index was created will be NOT be included. It will be included is the right thing to say and it's already explained in the sentence after that.