Fixed the contradicting sentence
Here the documentation states that copilot will not be able to edit specific excluded files that are configured. The next sentence then states it will “not” ignore them and it will have the ability to see and update them. I’ve fixed this contradiction so that now the documentation make sense.